requirements.txt
numpy>=2.1.3
pandas>=2.2.3
pyarrow>=18.0.0
python_dateutil==2.9.0.post0
PyYAML>=6.0.2
scikit_learn>=1.5.2
setuptools==75.6.0
torch==2.5.1
tqdm==4.67.1
transformers>=4.48.0
CausalEstimate==0.8.0
lightning>=2.4.0
xgboost>=3.0.0,<4.0.0
betacal==1.1.0